Make comm diags runtime start/stop/get implementation-independent.
The runtime code that starts, stops, and retrieves comm diags was
effectively the same for all comm layer implementations.  Make a single
copy of it, independent of the comm layers.
