+++ runtime/src/comm/ofi/comm-ofi.c
@@ -3001,77 +3001,6 @@ void chpl_comm_barrier(const char *msg) {
 
 ////////////////////////////////////////
 //
-// Comm diagnostics
-//
-
-void chpl_startVerboseComm() {
-  chpl_verbose_comm = 1;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(0 /* &chpl_verbose_comm */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_stopVerboseComm() {
-  chpl_verbose_comm = 0;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(0 /* &chpl_verbose_comm */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_startVerboseCommHere() {
-  chpl_verbose_comm = 1;
-}
-
-
-void chpl_stopVerboseCommHere() {
-  chpl_verbose_comm = 0;
-}
-
-
-void chpl_startCommDiagnostics() {
-  chpl_comm_diagnostics = 1;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(1 /* &chpl_comm_diagnostics */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_stopCommDiagnostics() {
-  chpl_comm_diagnostics = 0;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(1 /* &chpl_comm_diagnostics */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_startCommDiagnosticsHere() {
-  chpl_comm_diagnostics = 1;
-}
-
-
-void chpl_stopCommDiagnosticsHere() {
-  chpl_comm_diagnostics = 0;
-}
-
-
-void chpl_resetCommDiagnosticsHere() {
-  chpl_comm_diags_reset();
-}
-
-
-void chpl_getCommDiagnosticsHere(chpl_commDiagnostics *cd) {
-  chpl_comm_diags_copy(cd);
-}
-
-
-////////////////////////////////////////
-//
 // Time
 //
