+++ runtime/src/comm/ugni/comm-ugni.c
@@ -8327,85 +8327,6 @@ void local_yield(void)
 }
 
 
-void chpl_startVerboseComm()
-{
-  chpl_verbose_comm = 1;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(0 /* &chpl_verbose_comm */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_stopVerboseComm()
-{
-  chpl_verbose_comm = 0;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(0 /* &chpl_verbose_comm */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_startVerboseCommHere()
-{
-  chpl_verbose_comm = 1;
-}
-
-
-void chpl_stopVerboseCommHere()
-{
-  chpl_verbose_comm = 0;
-}
-
-
-void chpl_startCommDiagnostics()
-{
-  chpl_comm_diagnostics = 1;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(1 /* &chpl_comm_diagnostics */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_stopCommDiagnostics()
-{
-  chpl_comm_diagnostics = 0;
-  chpl_comm_diags_disable();
-  chpl_comm_broadcast_private(1 /* &chpl_comm_diagnostics */, sizeof(int),
-                              -1 /*typeIndex: unused*/);
-  chpl_comm_diags_enable();
-}
-
-
-void chpl_startCommDiagnosticsHere()
-{
-  chpl_comm_diagnostics = 1;
-}
-
-
-void chpl_stopCommDiagnosticsHere()
-{
-  chpl_comm_diagnostics = 0;
-}
-
-
-void chpl_resetCommDiagnosticsHere()
-{
-  chpl_comm_diags_reset();
-
-#define _PSZM(psv) PERFSTATS_STZ(psv);
-  PERFSTATS_DO_ALL(_PSZM);
-#undef _PSZM
-}
-
-
-void chpl_getCommDiagnosticsHere(chpl_commDiagnostics *cd)
-{
-  chpl_comm_diags_copy(cd);
-}
-
 void chpl_comm_ugni_help_register_global_var(int i, wide_ptr_t wide)
 {
 }
